requirements.txt
streamlit
python-dotenv

langchain==0.3.25
langchain-community==0.3.24
langchain-chroma==0.2.4
langchain-openai==0.3.16

chromadb
pypdf